[feat]: update users' current progress in race

+++ src/main/java/com/example/burningkey/multiplayer/service/UserDTO.java
@@ -16,7 +16,7 @@ import java.io.Serializable;
 public class UserDTO {
 
     private String username;
-    private Short currentWordPosition;
+    private Integer currentWordPosition;
     private String currentWord;
     private String sessionId;
     @JsonIgnore
